itsa: Allow fib_sleep() to be called subsequent times
fib_sleep() is used for doing exponential back off based on the Fibonaci
sequence.

Previously it used a static variable to keep track of the last number,
if you were to start the sequence again by passing in -1 then it would
just start returning numbers in the sequence from where it left off.

We currently only call this function from one place, but lets fix it so
that we can call it from multiple locations by using a 'state' variable
that is also passed into the function that is used to track the last
number instead.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>