itsa: Fix set_screens() to use actual screen size
Previously set_screens() just used the same mechanism as set_win_sz()
which gets the size of the terminal window.

Rather than doing that for the screen size, look in
/sys/class/graphics/fb0/virtual_size for it.

We can also use /sys/class/graphics/fb0/bits_per_pixel to get the colour
depth.

We still assume a scaling-factor of 1 and still only handle a single
screen, but should be good enough for now otherwise we'll likely need to
start doing things at the X/Wayland level.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>