itsa: Don't blindly error out in display_calculation_messages()
In display_calculation_messages() don't error out if there were no
messages and we got a 404 'Not Found' (NO_MESSAGES_PRESENT), this is an
OK situation.

We will still error out on a 404 'Not Found'
(MATCHING_RESOURCE_NOT_FOUND) as this means no calculation data was
found for the given NINO/calculation_id, which _may_ indicate a problem.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>