itsa: Fix stack corruption in print_bread_crumb()
With commit 6cf1cc5 ("src/Makefile: Compile with -fstack-protector") I'm
seeing the following crash

 pensionSavingsTaxCharges / pensionSavingsTaxChargesDetail / pensionSchemeUnauthorisedPayments / pensionSchemeUnauthorisedPaymen
                                         amount : 5000.99
                                        taxPaid : 5000.99
                                           rate : 20
                               chargeableAmount : 5000.99
*** stack smashing detected ***: terminated

A look at the coredump under gdb seemed to show it crashing in
print_json_tree() and indeed increasing the size of val[] fixed this.

However that wasn't the actual root of the problem, if you notice, the
last bread crumb entry is truncated.

The bread crumb is created in print_bread_crumb(), increasing the size
of str[] in there also fixes the crash but also fixes the truncation of
the bread crumbs.

We do actually overflow str[] earlier than the one it crashes on, but it
always crashes on this one.

Anyway, we increase the size of str[] in print_bread_crumb(), to fix
this issue and also we can actually decrease the size of val[] a little
so we aren't actually using any more stack space than before.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>