itsa: Add extra colour output infrastructure
While we currently have colourised output, there is no way to turn it
off.

This adds the main infrastructure that will allow that.

We could have just made the escape codes variables (that could then be
set to "") but then they couldn't be used as preprocessor string
literals and would have to have gone in as %s's in the format strings
would I don't think would have been the best from a readability
standpoint.

So the way this works now is with a simple parser that parses tags in a
format string that says how to colourise the text.

E.g to write 'Hello World' in blue

    printc("#BLUE#Hello World#RST#\n");

The tag name is between the #'s and the #RST# is for 'reset'.

printc() operates just like printf(3) and takes the same format strings.

There are also several helper functions (well, technically they are
macro wrappers around a generic function) for doing some common
printing, such as the error messages.

printic() - for INFO messages
printsC() - for OK messages
printcc() - for CONFIRM messages
printwc() - for WARNING messages
printec() - for ERROR messages

Colourised output can be disabled by setting the NO_COLOR environment
variable, its value (if any) is unimportant, which simply causes the
tags to be removed from the format string.

This follows the informal standard at https://no-color.org/

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>