itsa: Fix a crash with create-period
When trying to create a periodic update we were getting a segfault.

This was introduced by commit 498b322 ("itsa: Update for API changes")
which changed get_period() to use
mtd_ob_list_inc_and_expend_obligations() instead of
mtd_sa_se_list_obligations() to get the list of periods.

Unfortunately that isn't enough, as the format of the returned JSON has
totally changed which meant were trying to strdup(3) a NULL pointer.

We basically need to do the same changes we _did_ make to list_period().

This still doesn't actually work but that seems more due to the HMRC MTD
testing environment[0]:

[0]: https://github.com/hmrc/income-tax-mtd-changelog/issues/3

Fixes: 498b322 ("itsa: Update for API changes")
Closes: https://github.com/ac000/itsa/issues/2
Signed-off-by: Andrew Clayton <andrew@digital-domain.net>