color: Fix a va_end() snafu
In printc_xtra() we were calling va_end(args) before we had finished
with args and as the man page states, after va_end() the state of args
is undefined.

Now everything was working fine when compiled with GCC, however when
compiling with clang, we got a segfault the first time we tried using
printc_xtra().

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>