itsa: Preliminary support for multiple businesses
This is the start of supporting multiple businesses. When the 'init'
command is run, it will lookup all the users businesses and present them
to the user where they can then select the default business and provide
the path to the accounting data source for this business. This
'gnc_sqlite' is now a per business property rather than global.

We store the list of businesses in the config.json as an array with an
integer 'business_idx' value pointing into that array for the business
to operate on.

SEID (Self-Employment ID) has been renamed to BUSINESS_ID reflecting the
similar change in HMRC.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>