itsa: Fix overriding of NO_COLOR
With commit 1529c74 ("color: Switch to textus_coloris") we lost the
ability to override the NO_COLOR environment variable, which if set will
disable colourised output.

Commit 09513c7 ("textus_coloris: Update to latest version") means we can
once again override NO_COLOR.

Previously this was done via a config file option. That's been switched
to an environment variable (ITSA_COLOR) as the config file option
produced a catch-22 in that if we fail to read the config we will want
to produce a colourised error message, even if the config file would
have disabled it.

Doing it as an environment variable removes this albeit slight edge
case.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>