itsa: Update get_eop_obligations() for API changes
A few months ago there were some MTD endpoint API changes that were
handled in commit 498b322 ("itsa: Update for API changes").
Unfortunately it didn't handle one of the changes fully, leaving
get-end-of-period-statement-obligations / get_eop_obligations() in a
broken segfault producing state.

While this function was updated for the change in libmtdac endpoint
function being called, it failed to take into account that the JSON
returned by this new endpoint function is different to the previous
function.

There are more levels of arrays to go through before we get the data we
need and the field names have changed,

Fixes: 498b322 ("itsa: Update for API changes")
Signed-off-by: Andrew Clayton <andrew@digital-domain.net>