itsa: fix up error handling in set_business()
The error handling in set_business() was a little, err... broken.

Even if we got an error here we would still return 0, rather -1 (or some
non-zero value).

We also weren't checking the return value of set_business in
do_init_all().

So fix all that up and make some of the error more useful by prefixing
them with "set_business: ".

Fixes: 08e0a46 ("itsa: Preliminary support for multiple businesses")
Signed-off-by: Andrew Clayton <andrew@digital-domain.net>