sqlite: Use prepared statements with parameterised inputs
The LGTM[0] service pointed out that technically the SQL queries could
be vulnerable to SQL injection.

In this case the SQL input effectively comes from the user themselves
and thus they'd only be shooting themselves in the foot.

However it's probably good practice to just avoid that situation
altogether by using prepared statements with parameterised inputs.

We move the preparation of the statements outside the while loop as they
need only be prepared once and then re-executed with adjusted inputs.

[0]: https://lgtm.com/

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>