itsa: Fix a potential snprintf(3) overflow
It's quite common to use the following pattern

  len += snprintf(buf + len, sizeof(buf) - len, ...);

Ignoring the fact that snprintf can return -1, it can also return a
value larger than the buffer being written to as if the string is
truncated snprintf will return how many bytes _would_ have been written.

The simplest way to deal with this is to use strlen(3) on buf after each
snprintf to set the value of len.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>