itsa: Allow to specify start & end dates to create_period()
This makes it possible to specify the start & end dates for the
create-period operation. These are optional and will still default to
using the first "open" period.

It _may_ be possible to get into a situation where a period has been
missed and you need to choose the next open one, not sure.

However this will also help with testing where currently the HMRC MTD
sandbox environment is in a state of flux with some stateful and some
stateless endpoints and there is a mismatch[0] in the periods.

[0]: https://github.com/hmrc/income-tax-mtd-changelog/issues/3#issuecomment-897196464

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>