Setup GitHub Actions
This adds an initial GitHub Actions Workflow.

It currently does builds under CentOS 7 & 8 with GCC, Debian 11 with GCC
and Fedora 34 with GCC & clang. All x86-64.

It uses docker containers as GitHub only supports directly running on
Ubuntu (as far as Linux is concerned).

It will run the build tests on pull-requests and pushes to master.

It should only trigger on updates to; Makefile, src/ &
.github/workflows/

It adds -Werror to the CFLAGS to try and keep a warning free build.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>