itsa: Enhance the list_for_each() macro
list_for_each() is a simple macro for iterating over a singly linked
list, before calling it we take a copy of the list pointer so as not to
loose the start of the list and pass that into the macro.

This slightly enhances the macro so that we now pass a variable into the
macro that will act as the copy and set that in the macro's for loop.
This saves a line of code from not needing to make the copy ourselves.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>