Make use of the 'cleanup' variable __attribute__
A lot of the functions declare a char *jbuf which is malloc(3)'d by
libmtdac and which we free(3). This holds the response buffer.

Lets make use of the GCC/Clang 'cleanup' variable __attribute__ to set a
'free' function on these variables that will be automatically called
when they go out of scope.

The 'cleanup' function tales a pointer to a type compatible with the
variable in question. So in this case it gets a char **, so we just
write a simple wrapper around free(3) for this.

It should also be noted that it would often be the case with this
mechanism that you would need to initialise the pointer either to NULL
or with something like malloc(3) otherwise it may be uninitialised when
it goes out of scope.

In this case it's OK though, as it's immediately initialised to NULL in
mtd_ep() and is so is guaranteed to either be NULL or a valid pointer
after the call to mtd_ep() no matter what.

Signed-off-by: Andrew Clayton <ac@sigsegv.uk>